Rename SVGWriter interface
* rename SVGWriter to DefaultSVGWriter
* rename SVGWriterInterface to SVGWriter
