+++ substation-diagram/substation-diagram-core/src/main/java/com/powsybl/substationdiagram/VoltageLevelDiagram.java
@@ -16,8 +16,8 @@ import com.powsybl.substationdiagram.model.Graph;
 import com.powsybl.substationdiagram.svg.DefaultSubstationDiagramInitialValueProvider;
 import com.powsybl.substationdiagram.svg.DefaultSubstationDiagramStyleProvider;
 import com.powsybl.substationdiagram.svg.GraphMetadata;
+import com.powsybl.substationdiagram.svg.DefaultSVGWriter;
 import com.powsybl.substationdiagram.svg.SVGWriter;
-import com.powsybl.substationdiagram.svg.SVGWriterInterface;
 import com.powsybl.substationdiagram.svg.SubstationDiagramInitialValueProvider;
 import com.powsybl.substationdiagram.svg.SubstationDiagramStyleProvider;
 import org.slf4j.Logger;
@@ -62,15 +62,15 @@ public final class VoltageLevelDiagram {
     }
 
     public void writeSvg(ComponentLibrary componentLibrary, LayoutParameters layoutParameters, Network network, Path svgFile) {
-        SVGWriterInterface writer = new SVGWriter(componentLibrary, layoutParameters);
+        SVGWriter writer = new DefaultSVGWriter(componentLibrary, layoutParameters);
         writeSvg(writer, new DefaultSubstationDiagramInitialValueProvider(network), new DefaultSubstationDiagramStyleProvider(), svgFile, false);
     }
 
-    public void writeSvg(SVGWriterInterface writer, Network network, Path svgFile) {
+    public void writeSvg(SVGWriter writer, Network network, Path svgFile) {
         writeSvg(writer, new DefaultSubstationDiagramInitialValueProvider(network), new DefaultSubstationDiagramStyleProvider(), svgFile, false);
     }
 
-    public void writeSvg(SVGWriterInterface writer, SubstationDiagramInitialValueProvider initProvider, SubstationDiagramStyleProvider styleProvider, Path svgFile,
+    public void writeSvg(SVGWriter writer, SubstationDiagramInitialValueProvider initProvider, SubstationDiagramStyleProvider styleProvider, Path svgFile,
                          boolean debug) {
         Path dir = svgFile.toAbsolutePath().getParent();
         String svgFileName = svgFile.getFileName().toString();
@@ -89,11 +89,11 @@ public final class VoltageLevelDiagram {
                          SubstationDiagramInitialValueProvider initProvider, SubstationDiagramStyleProvider styleProvider,
                          Writer svgWriter,
                          Writer metadataWriter) {
-        SVGWriterInterface writer = new SVGWriter(componentLibrary, layoutParameters);
+        SVGWriter writer = new DefaultSVGWriter(componentLibrary, layoutParameters);
         writeSvg(writer, initProvider, styleProvider, svgWriter, metadataWriter);
     }
 
-    public void writeSvg(SVGWriterInterface writer, SubstationDiagramInitialValueProvider initProvider, SubstationDiagramStyleProvider styleProvider,
+    public void writeSvg(SVGWriter writer, SubstationDiagramInitialValueProvider initProvider, SubstationDiagramStyleProvider styleProvider,
             Writer svgWriter,
             Writer metadataWriter) {
         Objects.requireNonNull(writer);
