+++ substation-diagram/substation-diagram-core/src/test/java/com/powsybl/substationdiagram/TestVoltageLevelDiagram.java
@@ -14,8 +14,8 @@ import com.powsybl.substationdiagram.layout.LayoutParameters;
 import com.powsybl.substationdiagram.layout.PositionVoltageLevelLayoutFactory;
 import com.powsybl.substationdiagram.library.ComponentLibrary;
 import com.powsybl.substationdiagram.library.ResourcesComponentLibrary;
+import com.powsybl.substationdiagram.svg.DefaultSVGWriter;
 import com.powsybl.substationdiagram.svg.SVGWriter;
-import com.powsybl.substationdiagram.svg.SVGWriterInterface;
 import com.rte_france.powsybl.iidm.network.extensions.cvg.BusbarSectionPosition;
 import com.rte_france.powsybl.iidm.network.extensions.cvg.ConnectablePosition;
 import org.junit.Before;
@@ -88,7 +88,7 @@ public class TestVoltageLevelDiagram extends AbstractTestCase {
         ComponentLibrary componentLibrary = new ResourcesComponentLibrary("/ConvergenceLibrary");
         LayoutParameters layoutParameters = new LayoutParameters();
 
-        SVGWriterInterface writer = new SVGWriter(componentLibrary, layoutParameters);
+        SVGWriter writer = new DefaultSVGWriter(componentLibrary, layoutParameters);
 
         Path outSvg = tmpDir.resolve("vl.svg");
